api/main.py
"""
Physical AI & Humanoid Robotics Textbook - FastAPI Backend
RAG Chatbot + User Authentication + Personalization
"""

from fastapi import FastAPI, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, List
from datetime import datetime
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="Robotics Textbook API",
    description="AI-powered learning platform for robotics",
    version="0.1.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# ============================================================================
# Pydantic Models
# ============================================================================

class ChatQuery(BaseModel):
    """User chat query for RAG."""
    query: str
    user_id: Optional[str] = None
    context: Optional[str] = None

class ChatResponse(BaseModel):
    """Response from RAG chatbot."""
    answer: str
    sources: List[str]
    confidence: float
    timestamp: datetime

class UserProfile(BaseModel):
    """User profile data."""
    user_id: str
    email: str
    name: Optional[str] = None
    difficulty_level: str = "beginner"  # beginner, intermediate, advanced
    preferred_language: str = "en"  # en, ur
    modules_completed: List[int] = []
    created_at: datetime = None

class TranslationRequest(BaseModel):
    """Request for text translation."""
    text: str
    target_language: str = "ur"  # Urdu

class PersonalizationRequest(BaseModel):
    """Request for personalized content."""
    user_id: str
    chapter_id: str
    difficulty_level: Optional[str] = None

# ============================================================================
# Health Check
# ============================================================================

@app.get("/health")
async def health_check():
    """API health check."""
    return {
        "status": "healthy",
        "version": "0.1.0",
        "timestamp": datetime.now()
    }

# ============================================================================
# RAG Chatbot Endpoints
# ============================================================================

@app.post("/api/chat", response_model=ChatResponse)
async def chat(query: ChatQuery):
    """
    Chat with RAG-powered bot about robotics lessons.

    - Takes user query
    - Searches Qdrant vector DB for relevant lessons
    - Uses OpenAI to generate answer
    - Returns answer with citations
    """
    if not query.query:
        raise HTTPException(status_code=400, detail="Query cannot be empty")

    # TODO: Implement RAG pipeline
    # 1. Embed user query using OpenAI embeddings
    # 2. Search Qdrant for similar lesson chunks
    # 3. Retrieve top-3 sources
    # 4. Call OpenAI GPT with context
    # 5. Return answer + sources

    # Placeholder response
    return ChatResponse(
        answer="This is a placeholder response. RAG integration coming soon!",
        sources=["Module 1: Foundations", "Module 2: Programming"],
        confidence=0.0,
        timestamp=datetime.now()
    )

@app.post("/api/embed")
async def embed_lesson(lesson_id: str, content: str):
    """
    Embed lesson content into vector database.
    Called when new lessons are added or updated.

    - Splits content into chunks
    - Generates embeddings
    - Stores in Qdrant
    """
    # TODO: Implement embedding pipeline
    # 1. Split lesson into chunks
    # 2. Generate embeddings using OpenAI
    # 3. Store in Qdrant with metadata

    return {
        "status": "success",
        "lesson_id": lesson_id,
        "chunks_created": 0,
        "message": "Embedding pipeline coming soon"
    }

# ============================================================================
# Authentication Endpoints
# ============================================================================

@app.post("/api/auth/signup")
async def signup(email: str, password: str, name: Optional[str] = None):
    """
    User signup (Better-Auth integration).

    TODO: Connect to Better-Auth
    """
    # TODO: Implement signup
    # 1. Validate email
    # 2. Hash password
    # 3. Create user in Neon Postgres
    # 4. Generate session token

    return {
        "status": "success",
        "message": "Signup coming soon",
        "user": {
            "email": email,
            "name": name
        }
    }

@app.post("/api/auth/login")
async def login(email: str, password: str):
    """User login."""
    # TODO: Implement login
    return {
        "status": "success",
        "message": "Login coming soon",
        "token": "Bearer xxx"
    }

@app.get("/api/auth/profile")
async def get_profile(user_id: str):
    """Get user profile."""
    # TODO: Fetch from Neon Postgres
    return {
        "user_id": user_id,
        "email": "user@example.com",
        "name": "User Name",
        "difficulty_level": "beginner",
        "preferred_language": "en"
    }

# ============================================================================
# Personalization Endpoints
# ============================================================================

@app.post("/api/personalize/{user_id}/{chapter_id}")
async def personalize_content(user_id: str, chapter_id: str, request: PersonalizationRequest):
    """
    Get personalized content for a chapter.

    - Retrieves user preferences
    - Adapts content difficulty
    - Returns relevant lessons
    """
    # TODO: Implement personalization
    # 1. Fetch user profile
    # 2. Get chapter content
    # 3. Adapt based on difficulty level
    # 4. Return personalized content

    return {
        "status": "success",
        "chapter_id": chapter_id,
        "user_id": user_id,
        "content": "Personalized content coming soon"
    }

@app.post("/api/personalize/{user_id}/preferences")
async def update_preferences(user_id: str, preferences: dict):
    """Update user preferences."""
    # TODO: Update in Neon Postgres
    return {
        "status": "success",
        "message": "Preferences updated"
    }

# ============================================================================
# Translation Endpoints
# ============================================================================

@app.post("/api/translate")
async def translate(request: TranslationRequest):
    """
    Translate text to target language.
    Uses OpenAI API or cached translations.
    """
    # TODO: Implement translation
    # 1. Check cache in Neon for translation
    # 2. If not cached, use OpenAI translation API
    # 3. Cache result for future use
    # 4. Return translated text

    return {
        "original": request.text,
        "translated": "Translation coming soon",
        "target_language": request.target_language
    }

@app.get("/api/translate/{chapter_id}")
async def get_chapter_translation(chapter_id: str, language: str = "ur"):
    """Get full chapter translation."""
    # TODO: Fetch translation from Neon Postgres
    return {
        "chapter_id": chapter_id,
        "language": language,
        "content": "Full translation coming soon"
    }

# ============================================================================
# Progress Tracking
# ============================================================================

@app.post("/api/progress/{user_id}/{lesson_id}")
async def mark_lesson_complete(user_id: str, lesson_id: str):
    """Mark lesson as completed for user."""
    # TODO: Update progress in Neon
    return {
        "status": "success",
        "user_id": user_id,
        "lesson_id": lesson_id,
        "message": "Lesson marked as complete"
    }

@app.get("/api/progress/{user_id}")
async def get_user_progress(user_id: str):
    """Get user's overall progress."""
    # TODO: Calculate progress from Neon data
    return {
        "user_id": user_id,
        "lessons_completed": 0,
        "total_lessons": 12,
        "progress_percentage": 0.0,
        "modules": []
    }

# ============================================================================
# Admin Endpoints
# ============================================================================

@app.post("/api/admin/index-lessons")
async def index_all_lessons():
    """
    Admin endpoint to index all lessons into Qdrant.
    Call this when deploying to populate the vector DB.
    """
    # TODO: Implement lesson indexing
    # 1. Read all lesson markdown files
    # 2. Parse content
    # 3. Generate embeddings
    # 4. Store in Qdrant

    return {
        "status": "success",
        "lessons_indexed": 12,
        "message": "All lessons indexed successfully"
    }

# ============================================================================
# Error Handlers
# ============================================================================

@app.exception_handler(Exception)
async def global_exception_handler(request, exc):
    """Handle all exceptions."""
    return {
        "status": "error",
        "detail": str(exc),
        "timestamp": datetime.now()
    }

# ============================================================================
# Main
# ============================================================================

if __name__ == "__main__":
    import uvicorn

    uvicorn.run(
        app,
        host="0.0.0.0",
        port=8000,
        reload=True
    )
