api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
qdrant-client==2.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.0
aiofiles==23.2.1
python-multipart==0.0.6
cors==1.0.1
